The beginning of a new project by the simbolic ignore file
We all don't start with ignorance? Not knowing how the path may turn out, not knowing how difficult it can be? Isn't ignorance one of the main things for starting new adventures? Maybe if we knew all the sacrifices and stones in the future before starting it we would get bored and quit before the first step.

That's why I salute ignorance in this first commit with this Unity's based gitignore file.

May the journey beggin.